+++ Hearthstone Deck Tracker/Hearthstone/Watchers.cs
@@ -48,7 +48,7 @@ namespace Hearthstone_Deck_Tracker.Hearthstone
 			UiWatcher.Change += OnUiChange;
 			DeckPickerWatcher.Change += OnDeckPickerChange;
 			SceneWatcher.Change += (sender, args) => SceneHandler.OnSceneUpdate((Mode)args.PrevMode, (Mode)args.Mode, args.SceneLoaded, args.Transitioning);
-			ChoicesWatcher.Change += (sender, args) => Core.Overlay.SetChoicesVisible(args.CurrentChoice?.IsVisible ?? false, args.CurrentChoice?.Cards);
+			ChoicesWatcher.Change += (sender, args) => Core.Overlay.SetChoicesVisible(args.CurrentChoice?.IsVisible ?? false, args.CurrentChoice?.IsShopChoice ?? false, args.CurrentChoice?.Cards);
 			SpecialShopChoicesStateWatcher.Change += (sender, args) => Core.Overlay.HandleSpecialShop(args);
 			DiscoverStateWatcher.Change += OnDiscoverStateChange;
 			BigCardWatcher.Change += OnBigCardChange;